Reorganize norm module
