+++ src/pykeen/nn/__init__.py
@@ -2,9 +2,12 @@
 
 """PyKEEN internal "nn" module."""
 
-from . import init, norm
+from . import init
+from .emb import Embedding, EmbeddingSpecification, RepresentationModule
 
 __all__ = [
+    'Embedding',
+    'EmbeddingSpecification',
+    'RepresentationModule',
     'init',
-    'norm',
 ]
